Add additional valid SCSS tests
