+++ __tests__/block-opening-brace-space-before.test.mjs
@@ -50,3 +50,25 @@ describe('flags warnings with block-opening-brace-space-before lint', () => {
 		);
 	});
 });
+
+describe('does not flag warnings with valid block-opening-brace-space-before', () => {
+	const validScss = (
+`.test-selector { color: #fff; }`);
+
+	let result;
+
+	beforeEach(async () => {
+		result = await stylelint.lint({
+			code: validScss,
+			config,
+		});
+	});
+
+	it('did not error', () => {
+		assert.equal(result.errored, false);
+	});
+
+	it('does not flag warnings', () => {
+		assert.equal(result.results[0].warnings.length, 0);
+	});
+});
