+++ __tests__/declaration-bang-space-after.test.mjs
@@ -50,3 +50,28 @@ describe('flags warnings with declaration-bang-space-after lint', () => {
 		);
 	});
 });
+
+describe('does not flag warnings with valid declaration-bang-space-after', () => {
+	const validScss = (
+`.test-selector {
+  color: #000 !important;
+}
+`);
+
+	let result;
+
+	beforeEach(async () => {
+		result = await stylelint.lint({
+			code: validScss,
+			config,
+		});
+	});
+
+	it('did not error', () => {
+		assert.equal(result.errored, false);
+	});
+
+	it('does not flag warnings', () => {
+		assert.equal(result.results[0].warnings.length, 0);
+	});
+});
