+++ __tests__/declaration-bang-space-before.test.mjs
@@ -48,3 +48,28 @@ describe('flags warnings with declaration-bang-space-before lint', () => {
 		);
 	});
 });
+
+describe('does not flag warnings with valid declaration-bang-space-before', () => {
+	const validScss = (
+`.test-selector {
+  color: #000 !important;
+}
+`);
+
+	let result;
+
+	beforeEach(async () => {
+		result = await stylelint.lint({
+			code: validScss,
+			config,
+		});
+	});
+
+	it('did not error', () => {
+		assert.equal(result.errored, false);
+	});
+
+	it('does not flag warnings', () => {
+		assert.equal(result.results[0].warnings.length, 0);
+	});
+});
