+++ __tests__/declaration-block-single-line-max-declarations.test.mjs
@@ -48,3 +48,26 @@ describe('flags warnings with declaration-block-single-line-max-declarations lin
 		);
 	});
 });
+
+describe('does not flag warnings with valid declaration-block-single-line-max-declarations', () => {
+	const validScss = (
+`.test-selector { color: #000; }
+`);
+
+	let result;
+
+	beforeEach(async () => {
+		result = await stylelint.lint({
+			code: validScss,
+			config,
+		});
+	});
+
+	it('did not error', () => {
+		assert.equal(result.errored, false);
+	});
+
+	it('does not flag warnings', () => {
+		assert.equal(result.results[0].warnings.length, 0);
+	});
+});
