+++ __tests__/declaration-colon-space-after.test.mjs
@@ -47,3 +47,26 @@ describe('flags warnings with declaration-colon-space-after lint', () => {
 		);
 	});
 });
+
+describe('does not flag warnings with valid declaration-colon-space-after', () => {
+	const validScss = (
+`.test-selector { color: #fff; }
+`);
+
+	let result;
+
+	beforeEach(async () => {
+		result = await stylelint.lint({
+			code: validScss,
+			config,
+		});
+	});
+
+	it('did not error', () => {
+		assert.equal(result.errored, false);
+	});
+
+	it('does not flag warnings', () => {
+		assert.equal(result.results[0].warnings.length, 0);
+	});
+});
