+++ __tests__/declaration-colon-space-before.test.mjs
@@ -55,3 +55,26 @@ describe('flags warnings with declaration-colon-space-before lint', () => {
 		);
 	});
 });
+
+describe('does not flag warnings with valid declaration-colon-space-before', () => {
+	const validScss = (
+`.test-selector { color: #fff; }
+`);
+
+	let result;
+
+	beforeEach(async () => {
+		result = await stylelint.lint({
+			code: validScss,
+			config,
+		});
+	});
+
+	it('did not error', () => {
+		assert.equal(result.errored, false);
+	});
+
+	it('does not flag warnings', () => {
+		assert.equal(result.results[0].warnings.length, 0);
+	});
+});
