+++ __tests__/indentation.test.mjs
@@ -47,3 +47,28 @@ color: #fff;
 		);
 	});
 });
+
+describe('does not flag warnings with valid indentation', () => {
+	const validScss = (
+`.test-selector {
+  color: #000;
+}
+`);
+
+	let result;
+
+	beforeEach(async () => {
+		result = await stylelint.lint({
+			code: validScss,
+			config,
+		});
+	});
+
+	it('did not error', () => {
+		assert.equal(result.errored, false);
+	});
+
+	it('does not flag warnings', () => {
+		assert.equal(result.results[0].warnings.length, 0);
+	});
+});
