+++ __tests__/number-leading-zero.test.mjs
@@ -47,3 +47,28 @@ describe('flags warnings with number-leading-zero lint', () => {
 		);
 	});
 });
+
+describe('does not flag warnings with valid number-leading-zero', () => {
+	const validScss = (
+`.test-selector {
+  line-height: 0.5em;
+}
+`);
+
+	let result;
+
+	beforeEach(async () => {
+		result = await stylelint.lint({
+			code: validScss,
+			config,
+		});
+	});
+
+	it('did not error', () => {
+		assert.equal(result.errored, false);
+	});
+
+	it('does not flag warnings', () => {
+		assert.equal(result.results[0].warnings.length, 0);
+	});
+});
