+++ __tests__/number-no-trailing-zeros.test.mjs
@@ -47,3 +47,29 @@ describe('flags warnings with number-no-trailing-zeros lint', () => {
 		);
 	});
 });
+
+
+describe('does not flag warnings with valid number-no-trailing-zeros', () => {
+	const validScss = (
+`.test-selector {
+  top: 1px;
+}
+`);
+
+	let result;
+
+	beforeEach(async () => {
+		result = await stylelint.lint({
+			code: validScss,
+			config,
+		});
+	});
+
+	it('did not error', () => {
+		assert.equal(result.errored, false);
+	});
+
+	it('does not flag warnings', () => {
+		assert.equal(result.results[0].warnings.length, 0);
+	});
+});
