+++ __tests__/scss-at-function-pattern.test.mjs
@@ -47,3 +47,28 @@ describe('flags warnings with scss/at-function-pattern lint', () => {
 		);
 	});
 });
+
+describe('does not flag warnings with valid scss/at-function-pattern', () => {
+	const validScss = (
+`@function calculation-function($some-number, $another-number) {
+  @return $some-number + $another-number;
+}
+`);
+
+	let result;
+
+	beforeEach(async () => {
+		result = await stylelint.lint({
+			code: validScss,
+			config,
+		});
+	});
+
+	it('did not error', () => {
+		assert.equal(result.errored, false);
+	});
+
+	it('does not flag warnings', () => {
+		assert.equal(result.results[0].warnings.length, 0);
+	});
+});
