+++ __tests__/scss-dollar-variable-colon-space-before.test.mjs
@@ -46,3 +46,26 @@ describe('flags warnings with scss/dollar-variable-colon-space-before lint', () 
 		);
 	});
 });
+
+describe('does not flag warnings with valid scss/dollar-variable-colon-space-after', () => {
+	const validScss = (
+`a { $var: 10px;}
+`);
+
+	let result;
+
+	beforeEach(async () => {
+		result = await stylelint.lint({
+			code: validScss,
+			config,
+		});
+	});
+
+	it('did not error', () => {
+		assert.equal(result.errored, false);
+	});
+
+	it('does not flag warnings', () => {
+		assert.equal(result.results[0].warnings.length, 0);
+	});
+});
