+++ __tests__/scss-dollar-variable-pattern.test.mjs
@@ -45,3 +45,25 @@ describe('flags warnings with invalid scss/dollar-variable-pattern lint', () => 
 		);
 	});
 });
+
+describe('does not flag warnings with valid scss/dollar-variable-pattern lint', () => {
+	const validScss = (
+`$my-var: 10px;`);
+
+	let result;
+
+	beforeEach(async () => {
+		result = await stylelint.lint({
+			code: validScss,
+			config,
+		});
+	});
+
+	it('did not error', () => {
+		assert.equal(result.errored, false);
+	});
+
+	it('does not flag warnings', () => {
+		assert.equal(result.results[0].warnings.length, 0);
+	});
+});
