+++ __tests__/scss-selector-no-redundant-nesting-selector.test.mjs
@@ -49,3 +49,29 @@ describe('flags warnings with scss/selector-no-redundant-nesting-selector lint',
 		);
 	});
 });
+
+describe('does not flag warnings with valid scss/selector-no-redundant-nesting-selector lint', () => {
+	const validScss = (
+`.parentreference {
+  > .bar {
+    color: #f00;
+  }
+}`);
+
+	let result;
+
+	beforeEach(async () => {
+		result = await stylelint.lint({
+			code: validScss,
+			config,
+		});
+	});
+
+	it('did not error', () => {
+		assert.equal(result.errored, false);
+	});
+
+	it('does not flag warnings', () => {
+		assert.equal(result.results[0].warnings.length, 0);
+	});
+});
