+++ __tests__/selector-pseudo-element-colon-notation.test.mjs
@@ -53,3 +53,31 @@ p::hover {
 		);
 	});
 });
+
+describe('does not flag warnings with valid selector-pseudo-element-colon-notation lint', () => {
+	const validScss = (
+`a::before {
+  content: '>';
+}
+
+a:hover {
+  color: #f00;
+}`);
+
+	let result;
+
+	beforeEach(async () => {
+		result = await stylelint.lint({
+			code: validScss,
+			config,
+		});
+	});
+
+	it('did not error', () => {
+		assert.equal(result.errored, false);
+	});
+
+	it('does not flag warnings', () => {
+		assert.equal(result.results[0].warnings.length, 0);
+	});
+});
