+++ __tests__/shorthand-property-no-redundant-values.test.mjs
@@ -47,3 +47,28 @@ describe('flags warnings with shorthand-property-no-redundant-values lint', () =
 		);
 	});
 });
+
+describe('does not flag warnings with valid shorthand-property-no-redundant-values', () => {
+	const validScss = (
+`.shorthand {
+  margin: 1px;
+}
+`);
+
+	let result;
+
+	beforeEach(async () => {
+		result = await stylelint.lint({
+			code: validScss,
+			config,
+		});
+	});
+
+	it('did not error', () => {
+		assert.equal(result.errored, false);
+	});
+
+	it('does not flag warnings', () => {
+		assert.equal(result.results[0].warnings.length, 0);
+	});
+});
