+++ __tests__/string-quotes.test.mjs
@@ -49,3 +49,26 @@ describe('flags warnings with string-quotes lint', () => {
 		);
 	});
 });
+
+describe('does not flag warnings with valid string-quotes', () => {
+	const validScss = (
+`.test-selector { content: 'test'; }
+`);
+
+	let result;
+
+	beforeEach(async () => {
+		result = await stylelint.lint({
+			code: validScss,
+			config,
+		});
+	});
+
+	it('did not error', () => {
+		assert.equal(result.errored, false);
+	});
+
+	it('does not flag warnings', () => {
+		assert.equal(result.results[0].warnings.length, 0);
+	});
+});
